\name{SymTS-package}
\alias{SymTS-package}
\alias{SymTS}
\docType{package}
\title{
\packageTitle{SymTS}
}
\description{
\packageDescription{SymTS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SymTS}
\packageIndices{SymTS}
}
\author{
\packageAuthor{SymTS}

Maintainer: \packageMaintainer{SymTS}
}
\references{
M. Grabchak (2016). Tempered Stable Distributions: Stochastic Models for Multiscale Processes. Springer, Cham.

S. T. Rachev, Y. S. Kim, M. L. Bianchi, and F. J. Fabozzi (2011). Financial Models with Levy Processes and Volatility Clustering. Wiley, Chichester.

J. Rosinski (2007). Tempering stable processes. Stochastic Processes and Their Applications, 117(6):677-707.

G. Samorodnitsky and M. Taqqu (1994). Stable Non-Gaussian Random Processes: Stochastic Models with Infinite Variance. Chapman & Hall, Boca Raton.
}
\keyword{Symmetric Tempered Stable Distributions}
\keyword{Symmetric Stable Distributions}